Refactored some more folder structure
