+++ apps/web/components/admin/AdminAddMCQ.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useState } from "react";
 import { Card, CardDescription, CardTitle, Button, Input, Sheet, SheetContent, SheetTrigger } from "@repo/ui";
 import { MCQQuestion, Problem } from "@prisma/client";
-import { createMCQ, deleteMCQ, getAllMCQQuestion } from "../components/utils";
-import EditMCQ from "./mcq/EditMCQ";
+import { createMCQ, deleteMCQ, getAllMCQQuestion } from "../utils";
+import EditMCQ from "../mcq/EditMCQ";
 
 interface AdminAddMCQProps extends Problem {
   mcqQuestions: MCQQuestion[];
