+++ apps/web/components/admin/CompleteTrackCard.tsx
@@ -1,8 +1,8 @@
 import { Prisma, ProblemType } from "@prisma/client";
 import { useEffect, useState } from "react";
 import { Sheet, SheetContent, SheetDescription, SheetHeader, SheetTitle, SheetTrigger, Button } from "@repo/ui";
-import EditProblem from "./EditProblem";
-import { createTrack } from "./utils";
+import EditProblem from "../EditProblem";
+import { createTrack } from "../utils";
 
 interface CompleteProblemCard {
   id: string;
